Allow upper case tune names in URLS
This has a lot of impact:

Firstly, search is now case insensitive because we use the Postgres ilike
operator instead of like which provides this function.

Secondly, it means that tune titles in lists will be capitalised properly on
the client.  Previously the tune url fragment in all lower case was visible.

All tests have been changed to use properly capitalised tune titles.